requirements.txt
# Multi-Tool Unifié - Dépendances Python
# Compatible Windows 10/11, Python 3.8+
# by Gsx - Intégration & Consolidation

# Core dependencies
requests>=2.28.0
colorama>=0.4.6
pystyle>=2.9

# Discord related
discord.py>=2.0.0
discord_webhook>=1.0.0
aiohttp>=3.8.0
pypresence>=4.2.0
dhooks>=1.1.4

# Web scraping & automation
selenium>=4.0.0
beautifulsoup4>=4.11.0
cloudscraper>=1.2.60
pywebcopy>=7.0.0
fake_useragent>=1.1.0

# OSINT & Information gathering
phonenumbers>=8.13.0
python-whois>=0.8.0
whois>=0.9.27
ipwhois>=1.2.0
dnspython>=2.3.0
emailrep>=0.0.5
holehe>=1.61
instaloader>=4.9.0

# Security & Cryptography
cryptography>=39.0.0

# Image & QR Code
Pillow>=9.4.0
qrcode>=7.4.0

# GUI & Automation
PyAutoGUI>=0.9.53
pyperclip>=1.8.2
plyer>=2.1.0

# Utilities
uuid>=1.30
fade>=0.0.1
deep_translator>=1.9.0
blessed>=1.20.0
setuptools>=65.0.0

# Optional (for building executables)
pyinstaller>=5.7.0
